skip microbit tests

+++ tests/cypress/e2e/tutoriale.cy.ts
@@ -33,6 +33,10 @@ import {assertStateDefinition} from "../support/expression-test-support";
 describe("Tutorial Link Navigation", () => {
     // Must clear all local storage between tests to reset the state:
     beforeEach(() => {
+        if (Cypress.env("VUE_APP_MICROBIT") === "true") {
+            cy.log("Skipping test in microbit mode");
+            return; // Skip the test
+        }
         cy.clearLocalStorage();
         cy.visit("/",  {onBeforeLoad: (win) => {
             win.localStorage.clear();
@@ -55,6 +59,10 @@ describe("Tutorial Link Navigation", () => {
 
 describe("Tutorials Page", () => {
     beforeEach(() => {
+        if (Cypress.env("VUE_APP_MICROBIT") === "true") {
+            cy.log("Skipping test in microbit mode");
+            return; // Skip the test
+        }
         cy.clearLocalStorage();
         cy.visit("/editor/#/tutorials",  {onBeforeLoad: (win) => {
             win.localStorage.clear();
@@ -93,6 +101,10 @@ describe("Tutorials Page", () => {
 
 describe("Upload Tutorial", () => {
     beforeEach(() => {
+        if (Cypress.env("VUE_APP_MICROBIT") === "true") {
+            cy.log("Skipping test in microbit mode");
+            return; // Skip the test
+        }
         cy.clearLocalStorage();
         cy.visit("/editor/#/tutorials",  {onBeforeLoad: (win) => {
             win.localStorage.clear();
@@ -151,6 +163,10 @@ describe("Upload Tutorial", () => {
 
 describe("Complete tutorial", () => {
     beforeEach(() => {
+        if (Cypress.env("VUE_APP_MICROBIT") === "true") {
+            cy.log("Skipping test in microbit mode");
+            return; // Skip the test
+        }
         cy.visit("/editor/#/tutPrimes");
     });
     it("Goes to the primes tutorial page", () => {
